case classes <-> json w/ no boilerplate or reflection.
This is largely proof of concept to get the shapeless side of persisting
case classes working, but maybe we will choose to keep it.

This commit uses Shapeless to add implement spray-json's ".toJson" method,
allowing case classes to be converted without any other code added.

JsonFormats includes instances of the JsonFormat typeclass for HLists of
Shapeless Field values (the types of which have been tagged with their
original field names), and an implicit that transforms case classes
into those HLists.

There is a basic test demonstrating the conversion.