Move shapeless derivation to separate module.
The code is reorganized to separate the shapeless specific typeclass derivation
to its own separate package to lead the way for alternate implementations.

In the future, the shapeless functionality should be moved to a submodule that
builds its own artifact with the shapeless dependency, so that the shapeless
dependency can be removed from the core project.